Refactoring multimer perm alignment, temporary fixes before removing recycling dimension
